解决了读写模型中的bug

+++ src/main/java/hust/tools/hmm/io/ObjectFileHMMWriter.java
@@ -39,7 +39,7 @@ public class ObjectFileHMMWriter extends AbstractHMMWriter {
 
 	@Override
 	public void writeCount(int count) throws IOException {
-		dos.writeLong(count);
+		dos.writeInt(count);
 	}
 
 	@Override
